+++ src/TJoinCalculator.h
@@ -12,10 +12,10 @@ class TJoinCalculator {
 public:
     explicit TJoinCalculator(Graph const& baseGraph);
 
-    [[nodiscard]] TJoin get_minimum_zero_join(std::function<double(double)> const& cost_transform) const;
+    [[nodiscard]] TJoin get_minimum_zero_join(std::function<long(long)> const& cost_transform) const;
 
     [[nodiscard]] TJoin get_minimum_cost_t_join(
-            std::vector<NodeId> const& odd_nodes, std::function<double(double)> const& cost_transform
+            std::vector<NodeId> const& odd_nodes, std::function<long(long)> const& cost_transform
     ) const;
 
 private:
